moved stat chart to data display
