+++ src/components/dataDisplay/StatChart/StatsChartStyles.ts
@@ -1,9 +1,9 @@
 import MuiCardContent from '@mui/material/CardContent'
 import MuiCardActions from '@mui/material/CardActions'
 import { styled } from '@mui/material/styles'
-import { Card, CardHeader } from '../surfaces/Card'
-import Typography from '../dataDisplay/Typography'
-import { Gradient } from '../types'
+import Typography from '../Typography'
+import { Gradient } from 'components/types'
+import { Card, CardHeader } from 'components/surfaces/Card'
 
 export const StyledCard = styled(Card)(({ theme }) => ({
   display: 'inline-block',
