+++ src/stories/dataDisplay/StatChart/BarPreview.tsx
