+++ src/stories/dataDisplay/StatChart/ScatterPreview.tsx
