+++ src/stories/dataDisplay/StatChart/TypesPreview.tsx
