+++ src/stories/dataDisplay/StatChart/_mocks.ts
